robot/production/requirements.txt
cffi==1.15.0
gevent==21.12.0
greenlet==1.1.2
pycparser==2.21
rel==0.4.7
websocket-client==1.3.2
zope.event==4.5.0
zope.interface==5.4.0
RPi.GPIO